requirements.txt
#devops
black
click
pytest
pytest-cov
pylint
#rust based linter
ruff
boto3
nbval
nbqa

#web
fastapi
uvicorn
#datascience
jupyter
panda
numpy
polars
scikit-learn
matplotlib
seaborn
dask
